LayoutTests/svg/animations/multiple-begin-additive-animation-expected.txt

SVG 1.1 dynamic animation tests


This tests additive='sum' support on animate elements with multiple begin times

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS rect.x.animVal.value is 0
PASS rect.y.animVal.value is 100
PASS rect.x.animVal.value is 0
PASS rect.y.animVal.value is 166.67
PASS rect.x.animVal.value is 0
PASS rect.y.animVal.value is 366.60
PASS rect.x.animVal.value is 0
PASS rect.y.animVal.value is 366.73
PASS rect.x.animVal.value is 0
PASS rect.y.animVal.value is 499.93
PASS rect.x.animVal.value is 400
PASS rect.y.animVal.value is 500.06
PASS rect.x.animVal.value is 400
PASS rect.y.animVal.value is 566.67
PASS rect.x.animVal.value is 400
PASS rect.y.animVal.value is 633.33
PASS rect.x.animVal.value is 400
PASS rect.y.animVal.value is 633.33
PASS rect.x.animVal.value is 400
PASS rect.y.animVal.value is 700
PASS rect.x.animVal.value is 400
PASS rect.y.animVal.value is 766.60
PASS rect.x.animVal.value is 0
PASS rect.y.animVal.value is 766.67
PASS rect.x.animVal.value is 0
PASS rect.y.animVal.value is 833.33
PASS rect.x.animVal.value is 0
PASS rect.y.animVal.value is 900
PASS rect.x.animVal.value is 0
PASS rect.y.animVal.value is 900
PASS rect.x.animVal.value is 0
PASS rect.y.animVal.value is 900
PASS successfullyParsed is true

TEST COMPLETE
